refactor: util functions & split js code into modules
